requirements.txt
alabaster==0.7.12
anaconda-client==1.7.1
anaconda-project==0.8.2
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astroid==2.0.3
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.0
blinker==1.4
bokeh==0.13.0
boto==2.49.0
boto3==1.9.17
botocore==1.12.18
Bottleneck==1.2.1
bz2file==0.98
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
clyent==1.2.2
colorama==0.3.9
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3.1
cryptography-vectors==2.3.1
cycler==0.10.0
Cython==0.29
cytoolz==0.9.0.1
dask==0.19.3
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distributed==1.23.3
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
fbprophet==0.3.post2
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.6
gensim==3.4.0
gevent==1.3.6
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.13
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hupper==1.4.1
hyperlink==17.3.1
idna==2.7
imageio==2.3.0
imagesize==1.1.0
incremental==17.5.0
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.34.12
jupyterlab-launcher==0.13.1
keyring==15.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.25.0
locket==0.2.0
lxml==4.2.5
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
mpmath==1.0.0
msgpack==0.5.6
multipledispatch==0.6.0
nbconvert==5.4.0
nbformat==4.4.0
neo4j-driver==1.6.2
neotime==1.0.0
networkx==2.2
nltk==3.2.5
nose==1.3.7
notebook==5.7.0
numba==0.40.0
numexpr==2.6.6
numpy==1.15.2
numpydoc==0.8.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.46
openpyxl==2.5.8
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
partd==0.3.8
PasteDeploy==1.5.2
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
plaster==1.0
plaster-pastedeploy==0.6
plotly==3.4.1
pluggy==0.7.1
ply==3.11
prometheus-client==0.4.0
prompt-toolkit==2.0.5
psutil==5.4.7
ptyprocess==0.6.0
py==1.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pylint==2.1.1
PyMySQL==0.9.2
pyodbc==4.0.24
pyOpenSSL==18.0.0
pyparsing==2.2.2
PySocks==1.6.8
pystan==2.17.1.0
pytest==3.8.2
pytest-arraydiff==0.2
pytest-astropy==0.4.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-crfsuite==0.9.6
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.1
qtconsole==4.4.1
QtPy==1.5.1
repoze.lru==0.7
requests==2.19.1
requests-oauthlib==1.0.0
retrying==1.3.3
rope==0.10.7
ruamel-yaml==0.15.71
s3transfer==0.1.13
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.7.1
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.0.5
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
spyder==3.2.8
spyder-kernels==1.1.0
SQLAlchemy==1.2.12
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
translationstring==1.3
Twisted==17.5.0
twython==3.7.0
typed-ast==1.1.0
unicodecsv==0.14.1
urllib3==1.23
venusian==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.4
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.1.1
xlwings==0.12.1
xlwt==1.3.0
zict==0.1.3
zope.deprecation==4.3.0
zope.interface==4.5.0
